feat!: add a `direction` prop to `NavigationContainer` to specify rtl/ltr
BREAKING CHANGE: Previously the navigators tried to detect RTL automatically and
adjust the UI. However this is problematic since we cannot detect RTL in all cases
(e.g. on Web).

This adds an optional `direction` prop to `NavigationContainer` instead
so that user can specify when React Navigation's UI needs to be adjusted for RTL.